Deploy/ObjectDetection/Improved_DFFT/requirements.txt
addict==2.4.0
albumentations==1.3.0
#apex==0.1
appdirs==1.4.4
asynctest==0.13.0
attrs==22.2.0
#certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
cityscapesScripts==2.2.1
click==8.1.3
#codecov==2.1.12
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
coverage==7.1.0
cycler==0.11.0
Cython==0.29.33
easydict==1.10
exceptiongroup==1.1.0
filelock==3.9.0
flake8==5.0.4
fonttools==4.38.0
huggingface-hub==0.12.1
humanfriendly==10.0
idna==3.4
imagecorruptions==1.1.2
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
imutils==0.5.4
iniconfig==2.0.0
interrogate==1.5.0
isort==4.3.21
joblib==1.2.0
kiwisolver==1.4.4
kwarray==0.6.9
Markdown==3.4.1
markdown-it-py==2.2.0
matplotlib==3.7.0
mccabe==0.7.0
mdurl==0.1.2
mmcv-full==1.7.1
# Editable install with no version control (mmdet==2.11.0)
#-e /home/mayi/anaconda3/envs/dfft/lib/python3.8/site-packages/mmdet-2.11.0-py3.8.egg
mmlvis==10.5.3
mmpycocotools==12.0.3
model-index==0.1.11
msgpack==1.0.4
networkx==3.0
numpy==1.23.5
onnx==1.7.0
onnxruntime==1.5.1
opencv-contrib-python==4.8.0.74
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
openmim==0.3.6
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pluggy==1.0.0
protobuf==3.20.3
py==1.11.0
pycocotools==2.0.6
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.14.0
pyparsing==3.0.9
pyquaternion==0.9.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
rich==13.3.1
safetensors==0.2.8
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
tabulate==0.9.0
tensorboardX==2.6
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.8.13.dev0
toml==0.10.2
tomli==2.0.1
torch==1.13.0+cu117
torchaudio==0.13.0+rocm5.2
torchvision==0.14.0+cu117
tqdm==4.64.1
typing==3.7.4.3
typing_extensions==4.5.0
ubelt==1.2.3
urllib3==1.26.14
xdoctest==1.1.1
yacs==0.1.8
yapf==0.32.0
zipp==3.14.0
